requirements.txt
python-memcached
protobuf
